01-ImageClassifier/requirements.txt
tensorflow>=2.12.0
numpy>=1.23.5
pandas>=1.5.3
matplotlib>=3.7.1
streamlit>=1.22.0
Pillow>=9.5.0

jupyter